refactor(files): split files controller by route domain
